requirements.txt
asgiref==3.2.7
autopep8==1.5.2
certifi==2020.4.5.1
chardet==3.0.4
Django==2.2.4
django-ckeditor==5.9.0
django-crispy-forms==1.9.1
django-js-asset==1.2.2
django-timesince-shortener==2020.3.31
django-widget-tweaks==1.4.1
djangorestframework==3.7.3
gunicorn==20.0.4
idna==2.9
multitasking==0.0.9
numpy==1.18.4
pandas==1.0.3
Pillow==7.1.2
pycodestyle==2.6.0
python-dateutil==2.8.1
python-memcached==1.59
pytz==2020.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
stripe==2.50.0
urllib3==1.25.9
whitenoise==5.0.1
yfinance==0.1.54
